single java bean for detailworkflow and workflow
