+++ app/src/main/java/org/apache/taverna/mobile/data/model/Workflow.java
@@ -18,13 +18,18 @@
  */
 package org.apache.taverna.mobile.data.model;
 
+
 import org.simpleframework.xml.Attribute;
 import org.simpleframework.xml.Element;
+import org.simpleframework.xml.ElementList;
+import org.simpleframework.xml.Root;
 
 import android.os.Parcel;
 import android.os.Parcelable;
 
+import java.util.List;
 
+@Root(name = "workflow")
 public class Workflow implements Parcelable {
 
     @Attribute(name = "resource", required = false)
@@ -39,35 +44,83 @@ public class Workflow implements Parcelable {
     @Attribute(name = "version", required = false)
     String version;
 
+    @Element(name = "id", required = false)
+    private String elementId;
+
     @Element(name = "title")
     private String title;
+
+    @Element(name = "description", required = false)
+    String description;
 
     @Element(name = "type")
     private Type type;
 
-    @Element(name = "uploader")
+    @Element(name = "uploader", required = false)
     private Uploader uploader;
 
-    @Element(name = "created-at")
+    @Element(name = "created-at", required = false)
     private String createdAt;
 
-    @Element(name = "preview")
+    @Element(name = "updated-at", required = false)
+    private String updatedAt;
+
+    @Element(name = "preview", required = false)
     private String previewUri;
 
-    public String getResource() {
-        return resource;
+    @Element(name = "svg", required = false)
+    private String svgUri;
+
+    @Element(name = "license-type", required = false)
+    private LicenseType licenseType;
+
+    @Element(name = "content-uri", required = false)
+    String contentUri;
+
+    @Element(name = "content-type", required = false)
+    String contentType;
+
+    @ElementList(name = "tags", required = false)
+    List<Tag> tag;
+
+    public String getContentType() {
+        return contentType;
     }
 
-    public void setResource(String resource) {
-        this.resource = resource;
+    public void setContentType(String contentType) {
+        this.contentType = contentType;
     }
 
-    public String getUri() {
-        return uri;
+    public String getContentUri() {
+        return contentUri;
     }
 
-    public void setUri(String uri) {
-        this.uri = uri;
+    public void setContentUri(String contentUri) {
+        this.contentUri = contentUri;
+    }
+
+    public String getCreatedAt() {
+        return createdAt;
+    }
+
+    public void setCreatedAt(String createdAt) {
+        this.createdAt = createdAt;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public String getElementId() {
+        return elementId;
+    }
+
+    public void setElementId(String elementId) {
+        this.elementId = elementId;
     }
 
     public String getId() {
@@ -78,20 +131,44 @@ public class Workflow implements Parcelable {
         this.id = id;
     }
 
-    public String getVersion() {
-        return version;
+    public LicenseType getLicenseType() {
+        return licenseType;
     }
 
-    public void setVersion(String version) {
-        this.version = version;
+    public void setLicenseType(LicenseType licenseType) {
+        this.licenseType = licenseType;
     }
 
-    public String getTitle() {
-        return title;
+    public String getPreviewUri() {
+        return previewUri;
     }
 
-    public void setTitle(String title) {
-        this.title = title;
+    public void setPreviewUri(String previewUri) {
+        this.previewUri = previewUri;
+    }
+
+    public String getResource() {
+        return resource;
+    }
+
+    public void setResource(String resource) {
+        this.resource = resource;
+    }
+
+    public String getSvgUri() {
+        return svgUri;
+    }
+
+    public void setSvgUri(String svgUri) {
+        this.svgUri = svgUri;
+    }
+
+    public List<Tag> getTag() {
+        return tag;
+    }
+
+    public void setTag(List<Tag> tag) {
+        this.tag = tag;
     }
 
     public Type getType() {
@@ -110,22 +187,40 @@ public class Workflow implements Parcelable {
         this.uploader = uploader;
     }
 
-    public String getCreatedAt() {
-        return createdAt;
+    public String getUri() {
+        return uri;
     }
 
-    public void setCreatedAt(String createdAt) {
-        this.createdAt = createdAt;
+    public void setUri(String uri) {
+        this.uri = uri;
     }
 
-    public String getPreviewUri() {
-        return previewUri;
+    public String getVersion() {
+        return version;
     }
 
-    public void setPreviewUri(String previewUri) {
-        this.previewUri = previewUri;
+    public void setVersion(String version) {
+        this.version = version;
     }
 
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    public String getUpdatedAt() {
+        return updatedAt;
+    }
+
+    public void setUpdatedAt(String updatedAt) {
+        this.updatedAt = updatedAt;
+    }
+
+    public Workflow() {
+    }
 
     @Override
     public int describeContents() {
@@ -138,14 +233,19 @@ public class Workflow implements Parcelable {
         dest.writeString(this.uri);
         dest.writeString(this.id);
         dest.writeString(this.version);
+        dest.writeString(this.elementId);
         dest.writeString(this.title);
+        dest.writeString(this.description);
         dest.writeParcelable(this.type, flags);
         dest.writeParcelable(this.uploader, flags);
         dest.writeString(this.createdAt);
+        dest.writeString(this.updatedAt);
         dest.writeString(this.previewUri);
-    }
-
-    public Workflow() {
+        dest.writeString(this.svgUri);
+        dest.writeParcelable(this.licenseType, flags);
+        dest.writeString(this.contentUri);
+        dest.writeString(this.contentType);
+        dest.writeTypedList(this.tag);
     }
 
     protected Workflow(Parcel in) {
@@ -153,11 +253,19 @@ public class Workflow implements Parcelable {
         this.uri = in.readString();
         this.id = in.readString();
         this.version = in.readString();
+        this.elementId = in.readString();
         this.title = in.readString();
+        this.description = in.readString();
         this.type = in.readParcelable(Type.class.getClassLoader());
         this.uploader = in.readParcelable(Uploader.class.getClassLoader());
         this.createdAt = in.readString();
+        this.updatedAt = in.readString();
         this.previewUri = in.readString();
+        this.svgUri = in.readString();
+        this.licenseType = in.readParcelable(LicenseType.class.getClassLoader());
+        this.contentUri = in.readString();
+        this.contentType = in.readString();
+        this.tag = in.createTypedArrayList(Tag.CREATOR);
     }
 
     public static final Creator<Workflow> CREATOR = new Creator<Workflow>() {
