+++ app/src/main/java/org/apache/taverna/mobile/ui/workflowdetail/WorkflowDetailMvpView.java
@@ -1,7 +1,7 @@
 package org.apache.taverna.mobile.ui.workflowdetail;
 
 
-import org.apache.taverna.mobile.data.model.DetailWorkflow;
+import org.apache.taverna.mobile.data.model.Workflow;
 import org.apache.taverna.mobile.data.model.License;
 import org.apache.taverna.mobile.data.model.User;
 import org.apache.taverna.mobile.ui.base.MvpView;
@@ -10,7 +10,7 @@ public interface WorkflowDetailMvpView extends MvpView {
 
     void showProgressbar(boolean b);
 
-    void showWorkflowDetail(DetailWorkflow detailWorkflow);
+    void showWorkflowDetail(Workflow workflow);
 
     void setImage(User user);
